requirements.txt
flask==2.2.3
flask-dynamo==0.1.2
boto3==1.26.135
botocore==1.29.135
python-dotenv==1.0.0
werkzeug==2.2.3